Module cv::cuda

Bindings to OpenCV’s classes and functions that exploits GPU/Cuda. See cv::cuda

Structs

GpuCascade
Data structure that performs object detection with a cascade classifier.

GpuHog
Data structure that performs Histogram of Gradient (HOG).

GpuMat
GpuMat data structure in rust, bound to an opaque type in C/C++.

Enums

CGpuCascade
Opaque data struct for C bindings

CGpuHog
Opaque data struct for C bindings

CGpuMat
Opaque data struct for C/C++ cv::cuda::GpuMat bindings