Struct cv::features2d::DescriptorMatcher

pub struct DescriptorMatcher { /* fields omitted */ }

Type for matching keypoint descriptors

Methods

impl DescriptorMatcher

pub fn new(descriptor_matcher_type: DescriptorMatcherType) -> DescriptorMatcher

Creates a descriptor matcher of a given type with the default parameters (using default constructor).

pub fn add(&mut self, descriptors: &Vec<&Mat>)

Adds descriptors to train a CPU or GPU descriptor collection

pub fn train(&mut self)

Trains a descriptor matcher

pub fn is_empty(&self) -> bool

Returns true if there are no train descriptors

pub fn match_(&self, query_descriptors: &Mat) -> Vec<DMatch>

Finds the best match for each descriptor from a query set

pub fn match_two(
&self,
query_descriptors: &Mat,
train_descriptors: &Mat
) -> Vec<DMatch>

Finds the best match for each descriptor from a query set. Unlike match_, train descriptors collection are passed directly

pub fn knn_match(&self, query_descriptors: &Mat, k: usize) -> Vec<Vec<DMatch>>

Finds the k best matches for each descriptor from a query set.