Enum cv::imgcodecs::ImreadModes

pub enum ImreadModes {
    ImreadUnchanged,
    ImreadGrayscale,
    ImreadColor,
    ImreadAnydepth,
    ImreadAnycolor,
    ImreadLoadGdal,
    ImreadReducedGrayscale2,
    ImreadReducedColor2,
    ImreadReducedGrayscale4,
    ImreadReducedColor4,
    ImreadReducedGrayscale8,
    ImreadReducedColor8,
}

ImreadModes

Variants
ImreadUnchanged

If set, return the loaded image as is (with alpha channel, otherwise it gets cropped

ImreadGrayscale

If set, always convert image to the single channel grayscale image.

ImreadColor

If set, always convert image to the 3 channel BGR color image.

ImreadAnydepth

If set, return 16-bit/32-bit image when the input has the corresponding depth, otherwise convert it to 8-bit.

ImreadAnycolor

If set, the image is read in any possible color format.

ImreadLoadGdal

If set, use the gdal driver for loading the image.

ImreadReducedGrayscale2

If set, always convert image to the single channel grayscale image and the image size reduced 1/2.

ImreadReducedColor2

If set, always convert image to the 3 channel BGR color image and the image size reduced 1/2.

ImreadReducedGrayscale4

If set, always convert image to the single channel grayscale image and the image size reduced 1/4.

ImreadReducedColor4

If set, always convert image to the 3 channel BGR color image and the image size reduced 1/4.

ImreadReducedGrayscale8

If set, always convert image to the single channel grayscale image and the image size reduced 1/8.

ImreadReducedColor8

If set, always convert image to the 3 channel BGR color image and the image size reduced 1/8.