Enum cv::imgcodecs::ImwriteFlags

pub enum ImwriteFlags {
    ImwriteJpegQuality,
    ImwriteJpegProgressive,
    ImwriteJpegOptimize,
    ImwriteJpegRstInterval,
    ImwriteJpegLumaQuality,
    ImwriteJpegChromaQuality,
    ImwritePngCompression,
    ImwritePngStrategy,
    ImwritePngBilevel,
    ImwritePxmBinary,
    ImwriteWebpQuality,
    ImwritePamTupletype,
}

Imwrite flags

Variants
ImwriteJpegQuality

For JPEG, it can be a quality from 0 to 100 (the higher is the better). Default value is 95.

ImwriteJpegProgressive

Enable JPEG features, 0 or 1, default is False.

ImwriteJpegOptimize

Enable JPEG features, 0 or 1, default is False.

ImwriteJpegRstInterval

JPEG restart interval, 0 - 65535, default is 0 - no restart.

ImwriteJpegLumaQuality

Separate luma quality level, 0 - 100, default is 0 - don't use.

ImwriteJpegChromaQuality

Separate chroma quality level, 0 - 100, default is 0 - don't use.

ImwritePngCompression

For PNG, it can be the compression level from 0 to 9. A higher value means a smaller size and longer compression time. Default value is 3. Also strategy is changed to IMWRITE_PNG_STRATEGY_DEFAULT (Z_DEFAULT_STRATEGY).

ImwritePngStrategy

One of cv::ImwritePNGFlags, default is IMWRITE_PNG_STRATEGY_DEFAULT.

ImwritePngBilevel

Binary level PNG, 0 or 1, default is 0.

ImwritePxmBinary

For PPM, PGM, or PBM, it can be a binary format flag, 0 or 1. Default value is 1.

ImwriteWebpQuality

For WEBP, it can be a quality from 1 to 100 (the higher is the better). By default (without any parameter) and for quality above 100 the lossless compression is used.

ImwritePamTupletype

For PAM, sets the TUPLETYPE field to the corresponding string value that is defined for the format