Struct cv::objdetect::CascadeClassifier

pub struct CascadeClassifier { /* fields omitted */ }

Cascade classifier class for object detection.

Methods

impl CascadeClassifier

fn new() -> CascadeClassifier

Creates a cascade classifier, uninitialized. Before use, call load.

fn from_path<P: AsRef<Path>>(path: P) -> Result<Self>

Creates a cascade classifier using the model specified.

fn load<P: AsRef<Path>>(&self, path: P) -> Result<()>

Loads the classifier model from a path.

fn detect_multiscale(&self, mat: &Mat) -> Vec<Rect>

The default detection uses scale factor 1.1, minNeighbors 3, no min size or max size.

fn detect_with_params(
&self,
mat: &Mat,
scale_factor: f32,
min_neighbors: i32,
min_size: Size2i,
max_size: Size2i
) -> Vec<Rect>

Detects the object using parameters specified.

mat - Matrix of the type CV_8U containing an image where objects are detected.

scale_factor - Parameter specifying how much the image size is reduced at each image scale.

min_neighbors - Parameter specifying how many neighbors each candidate rectangle should have to retain it.

min_size - Minimum possible object size. Objects smaller than that are ignored.

max_size - Maximum possible object size. Objects larger than that are ignored

OpenCV has a parameter (flags) that's not used at all.