Struct cv::objdetect::HogParams

pub struct HogParams {
    pub win_size: Size2i,
    pub block_size: Size2i,
    pub block_stride: Size2i,
    pub cell_size: Size2i,
    pub nbins: i32,
    pub win_sigma: f64,
    pub l2hys_threshold: f64,
    pub gamma_correction: bool,
    pub nlevels: usize,
    pub hit_threshold: f64,
    pub win_stride: Size2i,
    pub padding: Size2i,
    pub scale: f64,
    pub group_threshold: i32,
}

Parameters that controls the behavior of HOG

Fields
win_size: Size2i

Detection window size. Align to block size and block stride

block_size: Size2i

Block size in pixels. Align to cell size. Only (16,16) is supported for now.

block_stride: Size2i

Block stride. It must be a multiple of cell size.

cell_size: Size2i

Cell size. Only (8, 8) is supported for now.

nbins: i32

Number of bins. Only 9 bins per cell are supported for now.

win_sigma: f64

Gaussian smoothing window parameter.

l2hys_threshold: f64

L2-Hys normalization method shrinkage.

gamma_correction: bool

Flag to specify whether the gamma correction preprocessing is required or not.

nlevels: usize

Maximum number of detection window increases.

hit_threshold: f64

Threshold for the distance between features and SVM classifying plane. Usually it is 0 and should be specfied in the detector coefficients (as the last free coefficient). But if the free coefficient is omitted (which is allowed), you can specify it manually here.

win_stride: Size2i

Window stride. It must be a multiple of block stride.

padding: Size2i

Padding

scale: f64

Coefficient of the detection window increase.

group_threshold: i32

Coefficient to regulate the similarity threshold. When detected, some objects can be covered by many rectangles. 0 means not to perform grouping.