Module cv::videoio

Media I/O, see OpenCV videoio

Structs

VideoCapture
Video capturing from video files, image sequences or cameras.

VideoWriter
VideoWriter provides easy access to write videos to files. - On Linux FFMPEG is used to write videos; - On Windows FFMPEG or VFW is used; - On MacOSX QTKit is used.

Enums

CapProp
Video capture's property identifier.

VideoWriterProperty
VideoWriter's property identifier.

Functions

codec_name
Converts from OpenCV's int to four character code.

fourcc
Converts from four character code to i32 for OpenCV.