Struct cv::videoio::VideoCapture

pub struct VideoCapture { /* fields omitted */ }

Video capturing from video files, image sequences or cameras.

Methods

impl VideoCapture

fn new(index: i32) -> Self

Creates a capture device with specified camera id. If there is a single camera connected, just pass 0.

fn from_path(path: &str) -> Self

Creates a capture device with the path of a video file (eg. video.avi). This also supports image sequence, eg. img_%02d.jpg, which will read samples like img_00.jpg, img_01.jpg, img_02.jpg, ...).

fn is_open(&self) -> bool

Returns true if video capturing has been initialized already.

fn read(&self) -> Option<Mat>

Grabs, decodes and returns the next video frame. read combines VideoCapture::grab and VideoCapture::retrieve in one call. This is the most convenient method for reading video files or capturing data from decode and return the just grabbed frame.

If no frames has been grabbed (camera has been disconnected, or there are no more frames in video file), the methods return None.

fn set(&self, property: CapProp, value: f64) -> bool

Sets a property in the VideoCapture.

fn get(&self, property: CapProp) -> Option<f64>

Gets a property in the VideoCapture.