Crate cv

This library primarily provides a binding and API for OpenCV 3.x.

This is a work-in-progress and modules/functions are implemented as needed. Attempts to use rust-bindgen or cpp_to_rust haven't been very successful (I probably haven't tried hard enough). There is another port opencv-rust which generates OpenCV bindings using a Python script.

Re-exports

pub use core::*;

pub use mat::*;

Modules

core
Core data structures in OpenCV

cuda
Bindings to OpenCV's classes and functions that exploits GPU/Cuda. See cv::cuda

errors
Errors for OpenCV bindings

features2d
Provide 2D image feature detectors and descriptor extractors

hash
The module brings implementations of different image hashing algorithms.

highgui
highgui: high-level GUI

imgcodecs
Image file reading and writing, see OpenCV imgcodecs.

imgproc
Image processing, see OpenCV imgproc.

mat
Mat

objdetect
Various object detection algorithms, such as Haar feature-based cascade classifier for object detection and histogram of oriented gradients (HOG).

text
Provides different algorithms for text detection and recognition in natural scene images

video
Video Analysis, see OpenCV video

videoio
Media I/O, see OpenCV videoio

Macros

path_to_cstring

string_to_cstring