Struct cv::cuda::GpuCascade

pub struct GpuCascade { /* fields omitted */ }

Data structure that performs object detection with a cascade classifier.

Methods

impl GpuCascade

pub fn from_path<P: AsRef<Path>>(path: P) -> Result<Self, Error>

Loads the classifier from a file.

Name of the file from which the classifier is loaded. Only the old haar classifier (trained by the haar training application) and NVIDIA's nvbin are supported for HAAR and only new type of OpenCV XML cascade supported for LBP. The working haar models can be found at opencv_folder/data/haarcascades_cuda/.

pub fn detect_multiscale(&self, mat: &GpuMat) -> Vec<Rect>

Detects objects of different sizes in the input image.

pub fn set_find_largest_object(&mut self, value: bool)

Sets whether or not to find the only largest object.

pub fn set_max_num_objects(&mut self, max: c_int)

Sets the maximum number of objects.

pub fn set_min_neighbors(&mut self, min: c_int)

Sets minimal neighbors required for a detection to be valid.

pub fn set_max_object_size(&mut self, max: Size2i)

Sets the maximun object size.

pub fn set_min_object_size(&mut self, min: Size2i)

Sets the minimal object size.

pub fn set_scale_factor(&mut self, factor: f64)

Sets the scale factor used in multiscale detection.

pub fn get_classifier_size(&self) -> Size2i

Returns the classifier size.

pub fn get_find_largest_object_flag(&self) -> bool

Returns if the CascadeClassifier will only return the largest object.

pub fn get_max_num_objects(&self) -> c_int

Returns the allowed maximal number of detected objects.

pub fn get_min_neighbors(&self) -> c_int

Returns the number of minimal neighbors required for a detection to be valid.

pub fn get_max_object_size(&self) -> Size2i

Returns the maximum object size.

pub fn get_min_object_size(&self) -> Size2i

Returns the minimal object size.

pub fn get_scale_factor(&self) -> f64

Returns the scale factor.