Module cv::highgui

highgui: high-level GUI

Enums

MouseEventTypes
Mouse Events

WindowFlags
Flags for highgui_named_window. This only supports a subset of all cv::WindowFlags because C/C++ allows enum with the same value but Rust is stricter.

Functions

highgui_destroy_window
Destroy the specified window with the given name.

highgui_named_window
Create a window that can be used as a placeholder for images and trackbars. All created windows are referred to by their names. If a window with the same name already exists, the function does nothing.

highgui_set_mouse_callback
Set mouse handler for the specified window (identified by name). A callback handler should be provided and optional user_data can be passed around.

Type Definitions

MouseCallback
Callback function for mouse events, primarily used in highgui_set_mouse_callback

MouseCallbackData
Pointer referring to the data used in MouseCallback