Struct cv::objdetect::HogParams

pub struct HogParams {
    pub win_size: Size2i,
    pub block_size: Size2i,
    pub block_stride: Size2i,
    pub cell_size: Size2i,
    pub nbins: c_int,
    pub win_sigma: f64,
    pub l2hys_threshold: f64,
    pub gamma_correction: bool,
    pub nlevels: usize,
    pub hit_threshold: f64,
    pub win_stride: Size2i,
    pub padding: Size2i,
    pub scale: f64,
    pub group_threshold: c_int,
    pub use_meanshift_grouping: bool,
    pub final_threshold: f64,
}

Parameters that controls the behavior of HOG.

Fields
win_size: Size2i

Detection window size. Align to block size and block stride. The default is 64x128, trained the same as original paper.

block_size: Size2i

Block size in pixels. Align to cell size. Only (16,16) is supported for now (at least for GPU).

block_stride: Size2i

Block stride. It must be a multiple of cell size.

cell_size: Size2i

Cell size. Only (8, 8) is supported for now.

nbins: c_int

Number of bins. Only 9 bins per cell are supported for now.

win_sigma: f64

Gaussian smoothing window parameter. Default -1 for CPU and 4.0 for GPU.

l2hys_threshold: f64

L2-Hys normalization method shrinkage. Default 0.2.

gamma_correction: bool

Flag to specify whether the gamma correction preprocessing is required or not. Default false.

nlevels: usize

Maximum number of detection window increases (HOG scales). Default: 64.

hit_threshold: f64

Threshold for the distance between features and SVM classifying plane. Usually it is 0 and should be specfied in the detector coefficients (as the last free coefficient). But if the free coefficient is omitted (which is allowed), you can specify it manually here.

win_stride: Size2i

Window stride. It must be a multiple of block stride.

padding: Size2i

Padding

scale: f64

Coefficient of the detection window increase.

group_threshold: c_int

Coefficient to regulate the similarity threshold. When detected, some objects can be covered by many rectangles. 0 means not to perform grouping.

use_meanshift_grouping: bool

The useMeanShiftGrouping parameter is a boolean indicating whether or not mean-shift grouping should be performed to handle potential overlapping bounding boxes. While this value should not be set and users should employ non-maxima suppression instead, we support setting it as a library function.

final_threshold: f64

The finalThreshold parameter is mainly used to select the clusters that have at least finalThreshold + 1 rectangles. This parameter is passed when meanShift is enabled; the function rejects the small clusters containing less than or equal to finalThreshold rectangles, computes the average rectangle size for the rest of the accepted clusters and adds those to the output rectangle list.