Struct cv::objdetect::SvmDetector

pub struct SvmDetector {
    pub inner: *mut CSvmDetector,
}

SvmDetector

Fields
inner: *mut CSvmDetector

Pointer to the inner data structure

Methods

impl SvmDetector

fn default_people_detector() -> SvmDetector

The built-in people detector.

The size of the default people detector is 64x128, that mean that the people you would want to detect have to be atleast 64x128.

fn daimler_people_detector() -> SvmDetector

Returns the Daimler people detector.