Struct cv::Mat

pub struct Mat {
    pub inner: *mut CMat,
    pub cols: i32,
    pub rows: i32,
    pub depth: i32,
}

This wraps OpenCV's Mat class which is designed for n-dimensional dense array. It's the most widely used data structure in image/video processing since images are often stored as Mat.

Fields
inner: *mut CMat

Pointer to the actual C/C++ data structure

cols: i32

Number of columns

rows: i32

Number of rows

depth: i32

Depth of this mat

Methods

impl Mat

fn from_raw(raw: *mut CMat) -> Mat

Creates a Mat object from raw CMat pointer. This will read the rows and cols of the image.

fn new() -> Mat

Creates an empty Mat struct.

fn with_size(rows: i32, cols: i32, t: i32) -> Self

Create an empty Mat with specific size (rows, cols and types).

fn from_path(path: &str, flags: i32) -> Self

Create a Mat from reading the image specified by the path.

fn data(&self) -> *const u8

Returns the raw data (as a uchar pointer)

fn total(&self) -> usize

Returns the total number of array elements. The method returns the number of array elements (a number of pixels if the array represents an image). For example, images with 1920x1080 resolution will return 2073600.

fn elem_size(&self) -> usize

Returns the matrix element size in bytes. The method returns the matrix element size in bytes. For example, if the matrix type is CV_16SC3 , the method returns 3*sizeof(short) or 6.

fn size(&self) -> Size2i

Returns the size of this matrix.

fn is_valid(&self) -> bool

Check if the Mat is valid or not.

fn roi(&self, rect: Rect) -> Mat

Return a region of interest from a Mat specfied by a Rect.

fn logic_and(&mut self, mask: Mat)

Apply a mask to myself.

fn flip(&mut self, code: FlipCode)

Flips an image around vertical, horizontal, or both axes.

fn show(&self, name: &str, delay: i32)

Calls out to highgui to show the image, the duration is specified by delay.

fn cv_type(&self) -> CvType

Returns the images type. For supported types, please see CvType.

impl Mat

fn in_range(&self, lowerb: Scalar, upperb: Scalar) -> Mat

Check if Mat elements lie between the elements of two other arrays (lowerb and upperb). The output Mat has the same size as self and CV_8U type.

fn mix_channels(&self,
nsrcs: isize,
ndsts: isize,
from_to: *const i32,
npairs: isize)
-> Mat

Copy specified channels from self to the specified channels of output Mat.

fn normalize(&self, alpha: f64, beta: f64, t: NormTypes) -> Mat

Normalize the Mat according to the normalization type.

impl Mat

fn rectangle(&self, rect: Rect)

Draws a simple, thick, or filled up-right rectangle.

fn rectangle_custom(&self,
rect: Rect,
color: Scalar,
thickness: i32,
linetype: LineTypes)

Draws a simple, thick, or filled up-right rectangle.

fn rectangle2f(&self, rect: Rect2f)

Draw a simple, thick, or filled up-right rectangle.

fn cvt_color(&self, code: ColorConversionCodes) -> Mat

Convert an image from one color space to another.

fn pyr_down(&self) -> Mat

Blurs an image and downsamples it. This function performs the downsampling step of the Gaussian pyramid construction.

fn resize_to(&self, dsize: Size2i, interpolation: InterpolationFlag) -> Mat

Resizes an image.

The function resize resizes the image down to or up to the specified size.

fn resize_by(&self, fx: f64, fy: f64, interpolation: InterpolationFlag) -> Mat

Resizes an image.

The function resize resizes the image down to or up to the specified size.

fn calc_hist(&self,
channels: *const c_int,
mask: Mat,
dims: c_int,
hist_size: *const c_int,
ranges: *const *const f32)
-> Mat

Calculate a histogram of an image.

fn calc_back_project(&self,
channels: *const i32,
hist: &Mat,
ranges: *const *const f32)
-> Mat

Calculate the back projection of a histogram. The function calculates the back project of the histogram.

impl Mat

fn imdecode(buf: &[u8], mode: ImreadModes) -> Mat

Decodes an image from buf according to the specified mode.

fn imencode(&self, ext: &str, f: Vec<ImwriteFlags>) -> Option<Vec<u8>>

Encodes an image; the encoding scheme depends on the extension provided; additional write flags can be passed in using a vector. If successful, returns an owned vector of the encoded image.

impl Mat

fn camshift(&self, wndw: Rect, criteria: &TermCriteria) -> RotatedRect

Finds an object center, size, and orientation; returns as RotatedRect.

wndw - initial search window.

criteria - stop criteria for the underlying meanShift.