Enum cv::videoio::CapProp

#[repr(C)]pub enum CapProp {
    PosMsec,
    PosFrames,
    PosAviRatio,
    FrameWidth,
    FrameHeight,
    Fps,
    Fourcc,
    FrameCount,
    Format,
    Mode,
    Brightness,
    Contrast,
    Saturation,
    Hue,
    Gain,
    Exposure,
    ConvertRgb,
    WhiteBalanceBlueU,
    Rectification,
    Monochrome,
    Sharpness,
    AutoExposure,
    Gamma,
    Temperature,
    Trigger,
    TriggerDelay,
    WhiteBalanceRedV,
    Zoom,
    Focus,
    Guid,
    IsoSpeed,
    Backlight,
    Pan,
    Tilt,
    Roll,
    Iris,
    Settings,
    Buffersize,
    Autofocus,
}

Video capture’s property identifier.

Variants

PosMsec

Current position of the video file in milliseconds or video capture timestamp.

PosFrames

0-based index of the frame to be decoded/captured next.

PosAviRatio

Relative position of the video file: 0 - start of the film, 1 - end of the film.

FrameWidth

Width of the frames in the video stream.

FrameHeight

Height of the frames in the video stream.

Fps

Frame rate.

Fourcc

4-character code of codec.

FrameCount

Number of frames in the video file.

Format

Format of the Mat objects returned by retrieve() .

Mode

Backend-specific value indicating the current capture mode.

Brightness

Brightness of the image (only for cameras).

Contrast

Contrast of the image (only for cameras).

Saturation

Saturation of the image (only for cameras).

Hue

Hue of the image (only for cameras).

Gain

Gain of the image (only for cameras).

Exposure

Exposure (only for cameras).

ConvertRgb

Boolean flags indicating whether images should be converted to RGB.

WhiteBalanceBlueU

Currently not supported

Rectification

Rectification flag for stereo cameras (note: only supported by DC1394 v 2.x backend currently)

Monochrome

Sharpness

AutoExposure

Gamma

Temperature

Trigger

TriggerDelay

WhiteBalanceRedV

Zoom

Focus

Guid

IsoSpeed

Backlight

Pan

Tilt

Roll

Iris

Settings

Buffersize

Autofocus